pyproject.toml
[tool.poetry]
name = "welhub"
version = "0.1.0"
description = ""
authors = ["GuilhermehVinicius <guilherme.ribeiro.business@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.27.1"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
